Rough implementation of --best
Just a simple polynomial ranking of contrast/size/distance.

Contrast is simply determined using a scaled laplace (as we don't pre-scale the
image: remember I want this to work on panoramas). In the beginning I was
thinking of scaling the distance using log-1, but it doesn't even seem
necessary for good results.

Factors were chosen by handwaving, using my photos as a reference.

A new -d/--debug switch was added, which super-imposes the factors next to the
detected faces for debugging purposes.